+++ src/biquad_bank.rs
@@ -1,221 +1,21 @@
 //! Bank of parallel biquad filters with SIMD acceleration.
 
-use core::marker::PhantomData;
-use core::ops::Neg;
-use hacker::Parameter;
-use wide::{f32x8, f64x4};
-
 use super::audionode::*;
-use super::prelude::{U4, U8};
+use super::biquad::*;
+use super::setting::*;
+use super::signal::*;
 use super::*;
-use crate::setting::Setting;
-use numeric_array::ArrayLength;
-
-pub trait Realx<Size: ArrayLength>: Num + Sized + Neg<Output = Self> {
-    const PI: Self;
-    const TAU: Self;
-    const SQRT_2: Self;
-    fn tan(self) -> Self;
-    fn exp(self) -> Self;
-    fn cos(self) -> Self;
-    fn sqrt(self) -> Self;
-    fn reduce_add(self) -> f32;
-    fn from_frame(frame: &Frame<f32, Size>) -> Self;
-    fn to_frame(self) -> Frame<f32, Size>;
-    fn set(&mut self, index: usize, value: f32);
-}
-
-impl Realx<U8> for f32x8 {
-    const PI: Self = f32x8::PI;
-    const TAU: Self = f32x8::TAU;
-    const SQRT_2: Self = f32x8::SQRT_2;
-
-    #[inline(always)]
-    fn tan(self) -> Self {
-        f32x8::tan(self)
-    }
-    #[inline(always)]
-    fn exp(self) -> Self {
-        f32x8::exp(self)
-    }
-    #[inline(always)]
-    fn cos(self) -> Self {
-        f32x8::cos(self)
-    }
-    #[inline(always)]
-    fn sqrt(self) -> Self {
-        f32x8::sqrt(self)
-    }
-    #[inline(always)]
-    fn reduce_add(self) -> f32 {
-        f32x8::reduce_add(self)
-    }
-    #[inline(always)]
-    fn from_frame(frame: &Frame<f32, U8>) -> Self {
-        f32x8::new((*frame.as_array()).into())
-    }
-    #[inline(always)]
-    fn to_frame(self) -> Frame<f32, U8> {
-        f32x8::to_array(self).into()
-    }
-    #[inline(always)]
-    fn set(&mut self, index: usize, value: f32) {
-        self.as_array_mut()[index] = value;
-    }
-}
-
-impl Realx<U4> for f64x4 {
-    const PI: Self = f64x4::PI;
-    const TAU: Self = f64x4::TAU;
-    const SQRT_2: Self = f64x4::SQRT_2;
-
-    #[inline(always)]
-    fn tan(self) -> Self {
-        f64x4::tan(self)
-    }
-    #[inline(always)]
-    fn exp(self) -> Self {
-        f64x4::exp(self)
-    }
-    #[inline(always)]
-    fn cos(self) -> Self {
-        f64x4::cos(self)
-    }
-    #[inline(always)]
-    fn sqrt(self) -> Self {
-        f64x4::sqrt(self)
-    }
-    #[inline(always)]
-    fn reduce_add(self) -> f32 {
-        f64x4::reduce_add(self) as f32
-    }
-    #[inline(always)]
-    fn from_frame(frame: &Frame<f32, U4>) -> Self {
-        let array = frame.as_array();
-        let f64_array = [
-            f64::from(array[0]),
-            f64::from(array[1]),
-            f64::from(array[2]),
-            f64::from(array[3]),
-        ];
-        f64x4::new(f64_array)
-    }
-    #[inline(always)]
-    fn to_frame(self) -> Frame<f32, U4> {
-        let array_f64: [f64; 4] = f64x4::to_array(self);
-        let array_f32: [f32; 4] = array_f64.map(|x| x as f32);
-        array_f32.into()
-    }
-    #[inline(always)]
-    fn set(&mut self, index: usize, value: f32) {
-        self.as_array_mut()[index] = value as f64;
-    }
-}
-
-/// BiquadBank coefficients in normalized form using SIMD.
-#[derive(Copy, Clone, Debug, Default)]
-pub struct BiquadCoefsBank<F, Size>
-where
-    F: Realx<Size>,
-    Size: ArrayLength,
-{
-    pub a1: F,
-    pub a2: F,
-    pub b0: F,
-    pub b1: F,
-    pub b2: F,
-    _marker: PhantomData<Size>,
-}
-
-impl<F, Size> BiquadCoefsBank<F, Size>
-where
-    F: Realx<Size>,
-    Size: ArrayLength,
-{
-    /// Return settings for a Butterworth lowpass filter-bank.
-    /// Sample rate is in Hz.
-    /// Cutoff is the -3 dB point of the filter in Hz.
-    #[inline]
-    pub fn butter_lowpass(sample_rate: f32, cutoff: F) -> Self {
-        let c = F::from_f64;
-        let sr = F::from_f32(sample_rate);
-        let f: F = (cutoff * F::PI / sr).tan();
-        let a0r: F = c(1.0) / (c(1.0) + F::SQRT_2 * f + f * f);
-        let a1: F = (c(2.0) * f * f - c(2.0)) * a0r;
-        let a2: F = (c(1.0) - F::SQRT_2 * f + f * f) * a0r;
-        let b0: F = f * f * a0r;
-        let b1: F = c(2.0) * b0;
-        let b2: F = b0;
-        Self {
-            a1,
-            a2,
-            b0,
-            b1,
-            b2,
-            _marker: PhantomData,
-        }
-    }
-
-    /// Return settings for a constant-gain bandpass resonator-bank.
-    /// Sample rate and center frequency are in Hz.
-    /// The overall gain of the filter is independent of bandwidth.
-    #[inline]
-    pub fn resonator(sample_rate: f32, center: F, q: F) -> Self {
-        let c = F::from_f64;
-        let sr = F::from_f32(sample_rate);
-        let r: F = (-F::PI * center / (q * sr)).exp();
-        let a1: F = c(-2.0) * r * (F::TAU * center / sr).cos();
-        let a2: F = r * r;
-        let b0: F = (c(1.0) - r * r).sqrt() * c(0.5);
-        let b1: F = c(0.0);
-        let b2: F = -b0;
-        Self {
-            a1,
-            a2,
-            b0,
-            b1,
-            b2,
-            _marker: PhantomData,
-        }
-    }
-
-    /// Arbitrary biquad.
-    #[inline]
-    pub fn arbitrary(a1: F, a2: F, b0: F, b1: F, b2: F) -> Self {
-        Self {
-            a1,
-            a2,
-            b0,
-            b1,
-            b2,
-            _marker: PhantomData,
-        }
-    }
-
-    ///// Frequency response at frequency `omega` expressed as fraction of sampling rate.
-    //pub fn response(&self, omega: f64) -> Complex64 {
-    //    let z1 = Complex64::from_polar(1.0, -f64::TAU * omega);
-    //    let z2 = z1 * z1;
-    //    /// Complex64 with real component `x` and imaginary component zero.
-    //    fn re<T: Float>(x: T) -> Complex64 {
-    //        Complex64::new(x.to_f64(), 0.0)
-    //    }
-    //    (re(self.b0) + re(self.b1) * z1 + re(self.b2) * z2)
-    //        / (re(1.0) + re(self.a1) * z1 + re(self.a2) * z2)
-    //}
-}
 
 /// 2nd order IIR filter bank implemented in normalized Direct Form I and SIMD.
-/// - Setting: coefficients as tuple Parameter::BiquadBank(a1, a2, b0, b1, b2).
-/// - Input 0: input signal.
-/// - Output 0: filtered signal.
+/// - Setting channel `i` coefficients: `Setting::biquad(a1, a2, b0, b1, b2).index(i)`.
+/// - Inputs: input signals.
+/// - Outputs: filtered signals.
 #[derive(Default, Clone)]
-pub struct BiquadBank<F, Size>
+pub struct BiquadBank<F>
 where
-    F: Realx<Size>,
-    Size: ArrayLength + Sync + Send,
+    F: Float,
 {
-    coefs: BiquadCoefsBank<F, Size>,
+    coefs: BiquadCoefs<F>,
     x1: F,
     x2: F,
     y1: F,
@@ -223,10 +23,9 @@ where
     sample_rate: f64,
 }
 
-impl<F, Size> BiquadBank<F, Size>
+impl<F> BiquadBank<F>
 where
-    F: Realx<Size>,
-    Size: ArrayLength + Sync + Send,
+    F: Float,
 {
     pub fn new() -> Self {
         Self {
@@ -235,7 +34,7 @@ where
         }
     }
 
-    pub fn with_coefs(coefs: BiquadCoefsBank<F, Size>) -> Self {
+    pub fn with_coefs(coefs: BiquadCoefs<F>) -> Self {
         Self {
             coefs,
             sample_rate: DEFAULT_SR,
@@ -243,23 +42,22 @@ where
         }
     }
 
-    pub fn coefs(&self) -> &BiquadCoefsBank<F, Size> {
+    pub fn coefs(&self) -> &BiquadCoefs<F> {
         &self.coefs
     }
 
-    pub fn set_coefs(&mut self, coefs: BiquadCoefsBank<F, Size>) {
+    pub fn set_coefs(&mut self, coefs: BiquadCoefs<F>) {
         self.coefs = coefs;
     }
 }
 
-impl<F, Size> AudioNode for BiquadBank<F, Size>
+impl<F> AudioNode for BiquadBank<F>
 where
-    F: Realx<Size>,
-    Size: ArrayLength + Sync + Send,
+    F: Float,
 {
-    const ID: u64 = 15;
-    type Inputs = Size;
-    type Outputs = Size;
+    const ID: u64 = 98;
+    type Inputs = F::Size;
+    type Outputs = F::Size;
 
     fn reset(&mut self) {
         self.x1 = F::zero();
@@ -286,24 +84,34 @@ where
     }
 
     fn set(&mut self, setting: Setting) {
-        if let Parameter::BiquadBank(index, a1, a2, b0, b1, b2) = setting.parameter() {
-            let mut coefs = self.coefs;
-            coefs.a1.set(*index, *a1);
-            coefs.a2.set(*index, *a2);
-            coefs.b0.set(*index, *b0);
-            coefs.b1.set(*index, *b1);
-            coefs.b2.set(*index, *b2);
+        if let Address::Index(index) = setting.direction() {
+            if let Parameter::Biquad(a1, a2, b0, b1, b2) = setting.parameter() {
+                self.coefs.a1.set(index, *a1);
+                self.coefs.a2.set(index, *a2);
+                self.coefs.b0.set(index, *b0);
+                self.coefs.b1.set(index, *b1);
+                self.coefs.b2.set(index, *b2);
+            }
         }
     }
 
-    //fn route(&mut self, input: &SignalFrame, frequency: f64) -> SignalFrame {
-    //    let mut output = SignalFrame::new(self.outputs());
-    //    output.set(
-    //        0,
-    //        input.at(0).filter(0.0, |r| {
-    //            r * self.coefs().response(frequency / self.sample_rate)
-    //        }),
-    //    );
-    //    output
-    //}
+    fn route(&mut self, input: &SignalFrame, frequency: f64) -> SignalFrame {
+        let mut output = SignalFrame::new(self.outputs());
+        for i in 0..self.outputs() {
+            let coefs = BiquadCoefs::<f32>::arbitrary(
+                self.coefs.a1.get(i),
+                self.coefs.a2.get(i),
+                self.coefs.b0.get(i),
+                self.coefs.b1.get(i),
+                self.coefs.b2.get(i),
+            );
+            output.set(
+                i,
+                input
+                    .at(i)
+                    .filter(0.0, |r| r * coefs.response(frequency / self.sample_rate)),
+            );
+        }
+        output
+    }
 }